+++ pyproject.toml
@@ -24,6 +24,7 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
 ]
 
+
 [tool.poetry.dependencies]
 python = "^3.11"
 pytest = "*"
@@ -48,6 +49,7 @@ ruff = {version = "*", optional = true}
 setuptools = "*"
 types-requests = {version = "^2.31.0.2", optional = true}
 
+
 [tool.poetry.extras]
 dev = [
     "pre-commit",
@@ -69,12 +71,12 @@ multi_line_output = 3
 include_trailing_comma = "True"
 use_parentheses = "True"
 combine_as_imports = "True"
-src_paths = "screenpy_examples"
-known_local_folder = "screenpy_examples"
+extend_skip = [".idea"]
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
 
 [tool.ruff]
 # Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
@@ -141,29 +143,34 @@ exclude = [
 # Assume Python 3.11.
 target-version = "py311"
 
+
 [tool.ruff.per-file-ignores]
 "__init__.py" = ["F401"]
-#"screenpy_examples/screenpy/readthedocs/resolutions/is_palpable.py" = ["E501"]
-#"screenpy_examples/screenpy_selenium/github/features/test_github_search.py" = ["E501"]
 "screenpy_examples/screenpy_selenium/the_internet/features_pytest/setup_selenium.py" = ["A003"]
+
 
 [tool.ruff.isort]
 combine-as-imports = true
 split-on-trailing-comma = true
 
+
 [tool.ruff.flake8-pytest-style]
 mark-parentheses = false
+
 
 [tool.ruff.pycodestyle]
 ignore-overlong-task-comments = true
 
+
 #[tool.ruff.flake8-annotations]
 #mypy-init-return = true
+
 
 [tool.ruff.flake8-annotations]
 # ignore returns types for functions that implicity or explicitly only return None
 suppress-none-returning = true
 allow-star-arg-any = true
+
 
 [tool.black]
 line-length = 88
