+++ screenpy_examples/screenpy/quietly_logging/actions/do_thing.py
@@ -2,9 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from screenpy.pacing import beat
-from screenpy.protocols import Performable
-from screenpy.resolutions import IsEqual
+from screenpy import IsEqual, Performable, beat
 
 from screenpy_examples.screenpy.quietly_logging.actions.see import See
 from screenpy_examples.screenpy.quietly_logging.questions import SimpleQuestion
