+++ screenpy_examples/screenpy/quietly_logging/actions/see_any_of.py
@@ -5,11 +5,7 @@ at least one of which is expected to be true.
 
 from typing import Tuple, Type, TypeVar, Union
 
-from screenpy import Actor
-from screenpy.exceptions import UnableToAct
-from screenpy.pacing import beat
-from screenpy.protocols import Answerable
-from screenpy.resolutions import BaseResolution
+from screenpy import Actor, Answerable, BaseResolution, UnableToAct, beat
 
 from .see import See
