+++ screenpy_examples/screenpy/quietly_logging/actions/tryto.py
@@ -2,8 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Tuple
 
-from screenpy.pacing import the_narrator
-from screenpy.protocols import Performable
+from screenpy import Performable, the_narrator
 
 if TYPE_CHECKING:
     from screenpy import Actor
