+++ screenpy_examples/screenpy/readthedocs/actions/pan.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from screenpy.pacing import beat
+from screenpy import beat
 
 from screenpy_examples.screenpy.readthedocs.abilities import ControlCameras
