+++ screenpy_examples/screenpy/readthedocs/actions/simultaneously.py
@@ -5,13 +5,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from screenpy.pacing import beat
+from screenpy import beat
 
 from screenpy_examples.screenpy.readthedocs.abilities import ControlCameras
 
 if TYPE_CHECKING:
-    from screenpy import Actor
-    from screenpy.protocols import Performable
+    from screenpy import Actor, Performable
 
 
 class Simultaneously:
