+++ screenpy_examples/screenpy/readthedocs/tasks/cut_to_closeup.py
@@ -2,8 +2,7 @@
 Cut to a closeup for some personal reactions.
 """
 
-from screenpy import Actor
-from screenpy.pacing import beat
+from screenpy import Actor, beat
 
 from screenpy_examples.screenpy.readthedocs.abilities import ControlCameras
 from screenpy_examples.screenpy.readthedocs.actions import JumpToCamera, Zoom
