+++ screenpy_examples/screenpy_playwright/swaglabs/questions/shopping_cart_badge_number.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from screenpy.pacing import beat
+from screenpy import beat
 from screenpy_playwright.questions import Text
 
 from screenpy_examples.screenpy_playwright.swaglabs.ui.header import CART_BADGE
