+++ screenpy_examples/screenpy_requests/httpbin/features/test_methods.py
@@ -4,9 +4,7 @@ API test examples that use all the HTTP methods.
 from __future__ import annotations
 
 import pytest
-from screenpy import Actor, then, when
-from screenpy.actions import See
-from screenpy.resolutions import IsEqualTo, ReadsExactly
+from screenpy import Actor, IsEqualTo, ReadsExactly, See, then, when
 from screenpy_requests.actions import SendAPIRequest, SendGETRequest
 from screenpy_requests.questions import (
     BodyOfTheLastResponse,
