+++ screenpy_examples/screenpy_selenium/github/features/test_github_search.py
@@ -23,8 +23,7 @@ from screenpy import (
     then,
     when,
 )
-from screenpy_selenium.abilities import BrowseTheWeb
-from screenpy_selenium.actions import Open, SaveScreenshot
+from screenpy_selenium import BrowseTheWeb, Open, SaveScreenshot
 
 from screenpy_examples.screenpy_selenium.github.tasks.search_github import SearchGitHub
 from screenpy_examples.screenpy_selenium.github.ui.github_home_page import URL
