+++ screenpy_examples/screenpy_selenium/github/questions/search_results_message.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from screenpy.pacing import beat
-from screenpy_selenium.questions import Text
+from screenpy import beat
+from screenpy_selenium import Text
 
 from screenpy_examples.screenpy_selenium.github.ui.github_search_results_page import (
     RESULTS_MESSAGE,
