+++ screenpy_examples/screenpy_selenium/the_internet/features_pytest/test_dropdowns.py
@@ -4,11 +4,8 @@ test structure. These tests exercise the selecting Actions.
 """
 from __future__ import annotations
 
-from screenpy.actions import See
-from screenpy.pacing import act, scene
-from screenpy.resolutions import ReadsExactly
-from screenpy_selenium.actions import Open, Select
-from screenpy_selenium.questions import Selected
+from screenpy import ReadsExactly, See, act, scene
+from screenpy_selenium import Open, Select, Selected
 
 from screenpy_examples.screenpy_selenium.the_internet.user_interface.dropdown import (
     THE_DROPDOWN,
