+++ screenpy_examples/screenpy_selenium/the_internet/features_pytest/test_key_presses.py
@@ -7,8 +7,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Callable, Tuple
 
 from screenpy import ReadsExactly, See, act, scene
-from screenpy_selenium.actions import Enter, Open, Wait
-from screenpy_selenium.questions import Text
+from screenpy_selenium import Enter, Open, Text, Wait
 
 from ..user_interface.key_presses import ENTRY_INPUT, RESULT_TEXT, URL
